### 2024-07-16 #6995 @editorialbot
[PRE REVIEW]: FaaSr: R Package for Function-as-a-Service Cloud Computing
**Submitting author:** @spark0510 (<a href="http://orcid.org/0009-0000-5357-804X">Sungjae Park</a>)
**Repository:** https://github.com/FaaSr/FaaSr-package
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.0
