### 2024-07-16 #6995 @danielskatz
In this case, specifically, please remove the ORCID line rather than using an invalid ORCID in the metadata
