### 2024-07-16 #6995 @renatof
Thank you @danielskatz - @spark0510 had been able create a GitHub Action to successfully generate a PDF of the paper, but it looks like we overlooked the ORCID and DOI issues. We'll get those addressed asap.
