+++ paper/paper.bib
@@ -42,7 +42,8 @@
   title={Numpywren: Serverless linear algebra},
   author={Shankar, Vaishaal and Krauth, Karl and Pu, Qifan and Jonas, Eric and Venkataraman, Shivaram and Stoica, Ion and Recht, Benjamin and Ragan-Kelley, Jonathan},
   journal={arXiv preprint arXiv:1810.09679},
-  year={2018}
+  year={2018},
+  doi={https://doi.org/10.48550/arXiv.1810.09679}
 }
 
 @inproceedings{Jonas17,
@@ -50,7 +51,8 @@
   author={Jonas, Eric and Pu, Qifan and Venkataraman, Shivaram and Stoica, Ion and Recht, Benjamin},
   booktitle={Proceedings of the 2017 symposium on cloud computing},
   pages={445--451},
-  year={2017}
+  year={2017},
+  doi={https://doi.org/10.1145/3127479.3128601}
 }
 
 @inproceedings{chard20,
@@ -58,7 +60,8 @@
   author={Chard, Ryan and Babuji, Yadu and Li, Zhuozhao and Skluzacek, Tyler and Woodard, Anna and Blaiszik, Ben and Foster, Ian and Chard, Kyle},
   booktitle={Proceedings of the 29th International symposium on high-performance parallel and distributed computing},
   pages={65--76},
-  year={2020}
+  year={2020},
+  doi={https://doi.org/10.1145/3369583.3392683}
 }
 
 @article{hostetter1997curl,
@@ -77,6 +80,7 @@
     year = {2023},
     note = {R package version 1.4.7},
     url = {https://CRAN.R-project.org/package=httr},
+    doi = {https://doi.org/10.32614/CRAN.package.httr}
   }
 
 @Manual{paws,
@@ -85,6 +89,7 @@
     year = {2023},
     note = {R package version 0.3.0},
     url = {https://CRAN.R-project.org/package=paws},
+    doi = {https://doi.org/10.32614/CRAN.package.paws}
   }
 
 @article{reznick1993using,
@@ -104,7 +109,8 @@
   booktitle={Proceedings of the 25th International Conference on World Wide Web},
   pages={263--273},
   year={2016},
-  organization={International World Wide Web Conferences Steering Committee}
+  organization={International World Wide Web Conferences Steering Committee},
+  doi={https://doi.org/10.1145/2872427.2883029}
 }
 
 @Manual{arrow,
@@ -113,6 +119,7 @@
   year = {2024},
   note = {R package version 16.1.0, https://arrow.apache.org/docs/r/},
   url = {https://github.com/apache/arrow/},
+  doi = {https://doi.org/10.32614/CRAN.package.arrow}
 }
 
 @article{lai2019evaluating,
@@ -123,7 +130,8 @@
   number={1},
   pages={e02567},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1002/ecs2.2567}
 }
 
 @article{Nust2020Rocker,
@@ -133,7 +141,8 @@
   volume={12},
   number={1},
   pages={437-461},
-  year={2020}
+  year={2020},
+  doi={https://doi.org/10.48550/arXiv.2001.10641}
 }
 
 @online{faasr-docker,
