### 2024-07-17 25df657
fixed dois by removing https URL
