### 2024-07-18 #6995 @renatof
Hi @danielskatz - my primary suggestion would be Tyson Swetnam (GitHub user tyson-swetnam) who is familiar with both the cloud/FaaS side and application side; I don't see him in the JOSS reviewer database and he may or may not have the cycles to review (but he could perhaps suggest someone in his team?)
Other suggestions based on the papers with similarity scores by the editorial bot and who are in the JOSS reviewer db would be:
gvegayon
nuest
equitable-equations
Thanks!
