### 2024-08-07 #7027 @tyson-swetnam
the repository LICENSE file is not compliant yet, it only states:

```
YEAR: 2024
COPYRIGHT HOLDER: FaaSr authors
```
