### 2024-08-07 #7027 @danielskatz
I think this is partly related to CRAN rules, where the full license is in [LICENSE.md](https://github.com/FaaSr/FaaSr-package/blob/main/LICENSE.md)
