### 2024-08-21 #7027 @yadudoc
@danielskatz The FaaSr repo points to a tutorial repo for instructions on running locally. I followed the instructions for local docker deploy with Rocker and got stuck. I'm going to give the Posit cloud alternative a try, hopefully by the weekend.
