### 2024-08-21 #7027 @tyson-swetnam
In the `FaaSr/FaaSr-tutorial` there are user's `"My_Minio_Bucket_ACCESS_KEY"` and `"My_MinioBucket_SECRET_KEY"` committed publicly. Should remove these and put "REPLACE_WITH_YOUR_MINIO_SECRET_KEY"

https://github.com/FaaSr/FaaSr-tutorial/blob/main/faasr_env
