### 2024-08-21 #7027 @tyson-swetnam
Open Storage Network is mentioned as potential storage, but are no tutorial instructions for setting this up (it would be highly similar to S3, I'm guessing).
