### 2024-08-21 #7027 @danielskatz
👋 @tyson-swetnam - thanks for putting in comments here. One way to track them so they don't get lost would to create issues in the source repo, and to mention `openjournals/joss-reviews#7027` in the, so that a little breadcrumb is created here that tracks if they are open or closed.
