### 2024-08-28 FaaSr/FaaSr-package#107 @yadudoc
Add a contributing guide
I'm reviewing this package as part of your JOSS submission. So far things look great, and I have a few items pending that I'm trying to figure out. I notice that the repo does not have a `CONTRIBUTING.md` document that github recommends. You might already have all this material in project documentation, even so, I would highly recommend adding this material to a dedicated `CONTRIBUTING.md` file.

Adding breadcrumbs: openjournals/joss-reviews#7027
