### 2024-08-28 FaaSr/FaaSr-package#108 @yadudoc
Error during workflow registration
I'm running into an error during the workflow registration step listed in the `FaaSr-tutorial` repo (https://github.com/FaaSr/FaaSr-tutorial?tab=readme-ov-file#register-and-invoke-the-simple-workflow-with-github-actions)

Here's a snippet of the error message I'm seeing while attempting to run the tutorial on Posit cloud:
```
> faasr_tutorial$register_workflow()

── Registering workflow github actions for repo: My_GitHub_Account ───────────────────────────────────────────
❓Enter repository visibility[private/public]
public
✔ Create github local directories
✔ Create github env files
✔ Create github payload file
✔ Create github REAME file
✔ Create github workflow yml file
✔ Create github local repository
✖ Error: Failed to create the repo
Error in faasr_register_workflow_git_remote_repo(token, response, private,  : 25% [1/4]   ETA:10s

FaaSr \ Registering workflow github actions ■■■■■■■■■■■■■■■■                  50% [2/4]   ETA: 4s
```

My first guess was that this might be an issue with my github PAT token. I went back and recreated the token with `workflow` and `read:org` scopes and loaded this up with `credentials::set_github_pat(force_new=TRUE)`. That did not fix the issue.
Please let me know if you have any recommendations for solving this. Thanks in advance!

This is part of my review efforts for your JOSS submission. Required for the functionality checklist item for openjournals/joss-reviews#7027
