### 2024-08-29 FaaSr/FaaSr-package#108 @renatof
hi @yadudoc - I just went through the posit tutorial again and verified that it worked for me following the tutorial page. I wonder if one of these could have happened:

1) perhaps you didn't update/save the faasr_env file with the new token?
2) perhaps the tutorial_simple.json wasn't configured/saved with a UserName that matched your GitHub account?

If it's not too much trouble for you, would it be possible for you to give it another go with a fresh Posit instance and the new token you created?

I sincerely appreciate your feedback and patience - thank you!

(ps: I haven't used credentials::set_github_pat(force_new=TRUE) before - not sure if there may be untested side-effects?)
