### 2024-08-29 FaaSr/FaaSr-package#108 @yadudoc
Hi @renatof Thanks for your response. I redid the whole process and I believe I got one step farther, but the setup did not complete successfully. Here's the log message:

```
> usethis::use_git_config(user.name = "yadudoc", user.email = "yadudoc1729@gmail.com")
> credentials::set_github_pat()
If prompted for GitHub credentials, enter your PAT in the password field
Using GITHUB_PAT from Yadu Nand Babuji (credential helper: cache)
> faasr_tutorial <- faasr(json_path="tutorial_simple.json", env="faasr_env")

── Start FaaSr client tools ───────────────────────────────────────────────────────────────────────────────────────────
✔ Successfully get configuration from tutorial_simple.json and faasr_env
✔ Create the FaaSr directory: faasr_gh_local_repo
✔ Create the FaaSr directory: faasr_data
✔ Ready to use FaaSr client tools:
  • $register_workflow
  • $invoke_workflow
  • $set_workflow_timer
  • $unset_workflow_timer
> faasr_tutorial$register_workflow()

── Registering workflow github actions for repo: My_GitHub_Account ────────────────────────────────────────────────────
❓Enter repository visibility[private/public]
public
✔ Create github local directories
✔ Create github env files
✔ Create github payload file
✔ Create github REAME file
✔ Create github workflow yml file
✔ Create github local repository
✔ Successfully create the repo
✖ Error: Failed to update the remote repository
Error in faasr_register_workflow_github_actions(faasr, cred) :                50% [2/4]   ETA: 5s
FaaSr | Registering workflow github actions ■■■■■■■■■■■■■■■■                  50% [2/4]   ETA: 6s
```

I'm pretty sure my github credentials are correct, because I do see this newly created repo from when I ran the last command I've pasted above. (https://github.com/yadudoc/FaaSr-tutorial). Is this a perms issue? Does the token need perms beyond the perms I've attached here:

Any ideas on what might be going wrong?
