### 2024-08-29 FaaSr/FaaSr-package#108 @renatof
Hi @yadudoc - thank you for the update.
It looks like your token has the right permissions.
I wonder if this is a side-effect of the first attempt; would you be able to delete the FaaSr-tutorial repo on github and retry (or configure a different repo name on "ActionRepoName": "FaaSr-tutorial") and retry?
Thanks!
