### 2024-08-30 6dfe30e
Added new CONTRIBUTING and CODE_OF_CONDUCT docs
