### 2024-08-30 c26b5ed
fixed links to description and code of conduct
