### 2024-08-30 FaaSr/FaaSr-package#107 @renatof
FYI @yadudoc I've added the CONTRIBUTING.md file to the main repo. thank you!
