### 2024-09-01 FaaSr/FaaSr-package#107 @yadudoc
Nicely done! I'm marking this issue as done.
