### 2024-09-01 FaaSr/FaaSr-package#108 @yadudoc
Hi @renatof, I tried changing the repo name and that didn't work either. I updated the `ActionRepoName`, re-registered the workflow and invoked it. I'm getting the same error as earlier:

```
> faasr_tutorial <- faasr(json_path="tutorial_simple.json", env="faasr_env")

── Start FaaSr client tools ───────────────────────────────────────────────────────────────────────────────────────────
✔ Successfully get configuration from tutorial_simple.json and faasr_env
✔ Ready to use FaaSr client tools:
  • $register_workflow
  • $invoke_workflow
  • $set_workflow_timer
  • $unset_workflow_timer
> faasr_tutorial$register_workflow()

── Registering workflow github actions for repo: My_GitHub_Account ────────────────────────────────────────────────────
❓Enter repository visibility[private/public]
public
✔ Create github local directories
✔ Create github env files
✔ Create github payload file
✔ Create github REAME file
✔ Create github workflow yml file
✔ Create github local repository
✔ Successfully create the repo
✖ Error: Failed to update the remote repository
Error in faasr_register_workflow_github_actions(faasr, cred) :                25% [1/4]   ETA: 9s
FaaSr \ Registering workflow github actions ■■■■■■■■■■■■■■■■                  50% [2/4]   ETA: 4s
```
