### 2024-09-01 #7027 @spark0510
Hi! @tyson-swetnam, thank you so much for your precious review!
The tutorial is using Minio "playground", which its keys are open to public and has no need to set things(permissions and so on) up. Therefore, as a part of the tutorial, we let users to use Minio playground.

Setting up S3, Minio, and Open storage network is introduced on the manual: https://faasr.io/_pages/credentials.html
Manuals are also included in the CRAN package.

We expected that users will be taking a look at manuals once they successfully complete the tutorial. Though, we can put the link on the tutorial!
