### 2024-09-01 #7027 @yadudoc
@danielskatz The repo and the project appear to be in great shape. Kudos to @spark0510 and I wanted to express my appreciation for all the help I've been getting from @renatof with debugging some issues with the tutorial 👍

I'm comfortable signing off on the documentation and the article. The last hold up is getting a minimal example/tutorial working on my end, and my debugging efforts are documented in this issue: https://github.com/FaaSr/FaaSr-package/issues/108. I'm waiting for this to be resolved to cross off **Functionality** and **Automated tests** from the checklist.
