### 2024-09-03 FaaSr/FaaSr-package#108 @renatof
Hello @yadudoc and thank you so much for your patience!
@spark0510  and I went through the whole tutorial again from scratch independently, and we were not able to reproduce the problem.

One thing we did notice is that the tutorial documentation did not explicitly state to choose the choice of a “Classic” token - that was a failure on my part, and I apologize for the lack of clarity; the tutorial has since been updated.

Maybe this could be the root cause of the problem? We have not yet tested with tokens other than classic tokens, and the error message you show happens when "git push" fails.

If you would be so kind to try again from scratch, I would suggest:

1. On Github, create a new short-lived classic token with workflow and read:org scopes
2. On Github, delete your FaaSr-tutorial repo
3. Start a new Posit cloud instance and set it up with FaaSr as per tutorial docs
4. Enter the token above in faasr_env and save it
5. Enter your GitHub user name in tutorial_simple,json and save it

Let us know if this works? Thank you again very much for your time and feedback, we really appreciate it!
