### 2024-09-07 FaaSr/FaaSr-package#109 @gvegayon
[JOSS review] gvegayon (draft)
THIS IS A WIP

- [ ] The list of authors includes Vahid Daneshmand and Quin Thomas, what was their contribution? I don't see any commits associated to them in the repo.
