### 2024-09-07 #7027 @gvegayon
Hi @spark0510, just finished my review [here](https://github.com/FaaSr/FaaSr-package/issues/109). The package is looking great! Documentation looks great and I love you have GHA for the CI up and running. Here are some points in which the package+submission could improve (details in the GitHub issue):

1. It is unclear what the contribution is of all authors. I can understand that sometimes coauthors don't make direct commits to the code, or they made other types of contributions (which is OK!) but I would like some clarification on that.
2. There are two other R packages on CRAN that can be used with AWS lambdas: `aws.lambda` and `lambdr`. I don't see those mentioned in the package. How do those compare to your project?
3. There is no enough code coverage of the project. This is a complicated issue as writing tests can take significant time. Currently, the package shows about 9% coverage. I would encourage you to increase that and at least ensure the core function(s) work as expected. I would recommend writing a GHA workflow in which you make a simple call using the core functionality of the package. You don't need 100%, but at least make sure the core functionality is tested properly.

There are other technical issues that I describe in my review (including a minor PR addressing some quick fixes). Happy to continue the discussion in the associated issue.
