### 2024-09-09 FaaSr/FaaSr-package#109 @danielskatz
Issues related to JOSS review https://github.com/openjournals/joss-reviews/issues/7027
