### 2024-09-09 #7027 @danielskatz
@yadudoc - I see a couple of unchecked items in your list. What do you need in order to check them off too?
