### 2024-09-09 FaaSr/FaaSr-package#108 @yadudoc
@renatof All of my latest tests were done with the classic token, so I don't think that is the issue. Do you know if the package drop logs somewhere that might have info to dig into?
