### 2024-09-09 #7027 @yadudoc
@danielskatz I have not been able to run the tutorial (https://github.com/FaaSr/FaaSr-tutorial) which I assumed was the easiest path to confirm the claimed functionality. That's the only hold-up at this point. I've had a few exchanges with the developers on this issue (https://github.com/FaaSr/FaaSr-package/issues/108) but I've not had much luck. I'm happy to check this off if any of the reviewers have been able to verify functionality.
