### 2024-09-09 FaaSr/FaaSr-package#109 @renatof
Thank you so much @gvegayon for the detailed comments!
We will work on these and add a more complete comment here, but in the meantime to address your first point: both Quinn and Vahid were instrumental to the specification of requirements and the design of FaaSr (Vahid also contributed to the FaaSr-Docker repo which is used in building Docker images for users).
