### 2024-09-10 FaaSr/FaaSr-package#108 @renatof
hi @yadudoc - thank you again for your help reviewing!

We do drop logs for actions in the S3 bucket, but the error you're seeing is at registration time before the actions run, so before any action logs are created.

The error message indicates that the repository failed to update, which means that the Github API request to create the repository was successful but the git CLI command used to push updates to the repo failed (this is LOC 440 in https://github.com/FaaSr/FaaSr-package/blob/main/R/faasr_client_api_github-actions.R).

This seems to me it would be an indication of an problem with the token, though it's odd that the repo was created but not updated. We haven't been able to reproduce the problem yet.

Looking at your token permissions screenshot in the thread above, just wanted to check if the gray boxes ticked automatically by github upon selection of "workflow" also looked the same for you?

![Screenshot 2024-09-10 at 3 19 22 PM](https://github.com/user-attachments/assets/91c7f0fb-fc6b-43c9-9688-9b7fe640a20b)

Thanks again!
