### 2024-09-17 #7027 @renatof
@danielskatz thanks for the ping - @spark0510 is currently on a trip overseas, I need to work with him on the great+detailed comment responses for @gvegayon - in summary, responding to the three key points raised above:
1. regarding author contributions, Quinn Thomas, Cayelan Carey and Vahid Daneshmand were instrumental to the specification of requirements as well as in the design of FaaSr. Vahid also contributed to the FaaSr-Docker repo which is used in building Docker images for users.
2. the lambdr and aws.lambda packages only work with AWS Lambda and do not support workflows/deployment across multiple FaaS providers (Lambda/OpenWhisk/GHA) under the same abstraction
3. good point, we acknowledge this needs to be improved and are discussing how to best address it
