### 2024-09-25 #7027 @danielskatz
👋 @renatof & @spark0510 - when do you think you will be able to work on the three key points and on getting the testing to work for @yadudoc ?
