### 2024-09-25 #7027 @renatof
hi @danielskatz - thanks for your patience, both myself and @spark0510 just moved and started new academic positions in different institutions and the beginning of the term is hitting us now.
I believe I covered two of the 3 key points above in the thread (contributions and differences w.r.t. other lambda packages). We hope to give a concrete timeline for addressing/implementing the suggestions by @gvegayon tomorrow.
Unfortunately we're still been unable to replicate the issue @yadudoc has brought up despite several tries - from the error message, it seems to be an issue with GitHub token, but could also be misconfiguration.
@yadudoc - would it be possible for you to give it one more go starting from scratch (with a new token, new posit session, deleting repos automatically created in previous sessions)?
Thanks!
