### 2024-09-26 FaaSr/FaaSr-package#108 @yadudoc
@renatof I made progress after redoing the steps from scratch. The workflow invocation worked, but when checking for outputs, it appears to have failed. Here's a snippet from the shell:

```
☰ Successfully registered all github actions
> faasr_tutorial$invoke_workflow()
✔ faasr_register_workflow_github_invoke: GitHub Action: Successfully invoked:start
> mc_ls("play/faasr/tutorial")
Error in `processx::run(binary, args, ...)`:
! System command 'mc' failed
---
Exit status: 1
Stderr: <empty>
---
Type .Last.error to see the more details.
> mc_ls("play/faasr/tutorial")
Error in `processx::run(binary, args, ...)`:
! System command 'mc' failed
---
Exit status: 1
Stderr: <empty>
---
Type .Last.error to see the more details.
> mc_cat("play/faasr/tutorial/sample1.csv")
Error in `processx::run(binary, args, ...)`:
! System command 'mc' failed
---
Exit status: 1
Stderr:
mc: <ERROR> Unable to read from `play/faasr/tutorial/sample1.csv`. Bucket `faasr` does not exist.
---
Type .Last.error to see the more details.
```
