### 2024-09-26 FaaSr/FaaSr-package#108 @renatof
Hi @yadudoc - glad to hear you got to this point, and thank you so much for trying again!

I think what happened is that the S3 faasr bucket we use in this tutorial is automatically deleted every so often. It's a free testing environment provided by minio; we create the test faasr bucket in the tutorial script but it may have been deleted by the time you checked.
If you can try to invoke the workflow again, but first create the bucket to ensure it's in the play server:
mc_mb('play/faasr')
faasr_tutorial$invoke_workflow()
mc_ls("play/faasr/tutorial")
