### 2024-09-27 FaaSr/FaaSr-package#111 @renatof
Improve code test coverage
The code needs more extensive test coverage - currently, running covr::package_coverage(".") yields approximately 9% coverage
