### 2024-09-27 FaaSr/FaaSr-package#112 @renatof
Make it explicit that some functions return nothing
Many functions return nothing, but that is not explicit. Add invisible(NULL) statements to them to make "nothing returned" explicit
