### 2024-09-27 FaaSr/FaaSr-package#113 @renatof
Implement a github action test for a single workflow action
Implement a workflow action that can be triggered manually with a JSON configuration using a public S3 bucket for testing. This action should exercise the core FaaSr functions intended for integration testing
