### 2024-09-27 FaaSr/FaaSr-package#114 @renatof
Replace paste0() with file.path() for path handling
Create paths using file.path() to enhance cross-platform compatibility
