### 2024-09-27 FaaSr/FaaSr-package#108 @yadudoc
It worked! Some steps took a bit of delay before the outputs popped up, I would recommend mentioning that in the tutorial.

Here's what I see now:
```
> mc_mb('play/faasr')
Bucket created successfully `play/faasr`.

> faasr_tutorial$invoke_workflow()
✔ faasr_register_workflow_github_invoke: GitHub Action: Successfully invoked:start
> mc_ls("play/faasr/tutorial")
data frame with 0 columns and 0 rows
> mc_ls("play/faasr/tutorial")
data frame with 0 columns and 0 rows
> mc_ls("play/faasr/tutorial")
data frame with 0 columns and 0 rows
> mc_cat("play/faasr/tutorial/sample1.csv")
Error in `processx::run(binary, args, ...)`:
! System command 'mc' failed
---
Exit status: 1
Stderr:
mc: <ERROR> Unable to read from `play/faasr/tutorial/sample1.csv`. Object does not exist.
---
Type .Last.error to see the more details.
> mc_ls("play/faasr/tutorial")
[1] "sample1.csv" "sample2.csv"
> mc_cat("play/faasr/tutorial/sample1.csv")
[1] "\"v1\",\"v2\",\"v3\"\n1,1,1\n2,4,8\n3,9,27\n4,16,64\n5,25,125\n6,36,216\n7,49,343\n8,64,512\n9,81,729\n10,100,1000\n"
```
