### 2024-09-27 #7027 @danielskatz
At this point, it seems like the issues blocking acceptance are those raised by @gvegayon, which @renatof and @spark0510 are working on, correct?
