### 2024-09-30 #7027 @renatof
@danielskatz we've consolidated our [replies] in response to @gvegayon 's [comments] here, and created a few issues to be addressed in the process - thanks again for all the suggestions, @gvegayon !

[comment 1] Authors.: The list of authors includes Vahid Daneshmand and Quin Thomas; what was their contribution? I don't see any commits associated with them in the repo.

[reply 1] regarding author contributions, Quinn Thomas, Cayelan Carey and Vahid Daneshmand were instrumental to the specification of requirements as well as in the design of FaaSr. Vahid also contributed to the FaaSr-Docker repo which is used in building Docker images for users.

[comment 2] Related packages. Doing a search of "AWS Lambdas" on CRAN yielded two other R packages that I failed to see mentioned in the paper: lambdr and aws.lambda. How these packages compare to FaaSr? The authors should mention these in their paper.

[reply 2] the lambdr and aws.lambda packages only work with AWS Lambda and do not support workflows/deployment across multiple FaaS providers (Lambda/OpenWhisk/GHA) under the same abstraction

[comment 3] Code coverage. Running covr::package_coverage(".") yields approximately 9% coverage (see below). Although I understand writing tests can be cumbersome, I would encourage you to add the coverage badge and ensure that all new commits should have a corresponding test as you continue extending/working on the R package. Again, you can defer this for after publication, but make sure you have an issue for this point (if you don't have one already).

[reply 3] thank you, and point well taken. we’ve created an issue for this - https://github.com/FaaSr/FaaSr-package/issues/111 - and it’s something we’re actively looking for a developer to help with

[comment 4] Many functions return "nothing", yet I see that that's not the explicit behavior, and sometimes, may be returning something; for instance, the following code returns the value 10, even when we are not doing so explicitly:

myfun0 <- \(x) return(x)

myfun1 <- \(x) {
  z <- myfun0(x)
}

x <- myfun1(10)
x # 10
You can use invisible(NULL) to make "nothing returned" explicit.

[reply 4] thank you for the suggestion! we will use invisible(NULL) to make "nothing returned" explicit - https://github.com/FaaSr/FaaSr-package/issues/112

[comment 5] Missing a test via GHA. I think the package would greatly improve if the authors included a GitHub Action in which they use the package via GHA/AWS. They could have a GHA that is triggered manually (or automatically) to check the basic functionality works.

[reply 5] thank you for the suggestion. a key challenge for fully automating end-to-end workflow tests here is dealing with credentials (both S3 and FaaS provider). we can make progress here by implementing a test with single workflow action that uses a public S3 bucket such as minio play - https://github.com/FaaSr/FaaSr-package/issues/113

[comment 6]  Improper way to call stop(). Across the package, the function stop() is called right after printing a message on the screen. The proper way to do that is to put the message as part of the stop() argument, e.g., instead of doing the following:

msg <- ....
message(msg)
stop()
do

stop(...)

[reply 6] thank you for the suggestion. we will check with CRAN about this - in our first submission to the package, we were asked to use the message function.

[comment 7] Improper way of using on.exit.: The package incorrectly uses on.exit(). Currently, it is not inserting any bugs, but most of the time, the authors call on.exit to restore the working directory while doing so again right before returning the function. The following is an example of what the package currently does. The best practice would be leave the setwd (or whatever other operations need to be done at the end) by the on.exit() call:

my_setwd <- \(..., which.) {
  message(which.)
  setwd(...)
}

myfun <- function(x) {
  wd <- getwd()
  on.exit(my_setwd(wd, which. = "Last"))
  my_setwd(file.path(wd, ".."), which. = "First")

  # This line shouldn't exist since `on.exit()` already does this
  my_setwd(wd, which. = "Second")
  return(x)

}

myfun(1)
#> First
#> Second
#> Last

[reply 7] Thank you for the suggestion! We received feedback upon submitting to CRAN, where they requested FaaSr to have on.exit() if any function is changing working directories. We have gone through the code and could not find an instance of using setwd() twice in one function, but we will do a more thorough review and remove a second setwd() if found

[comment 8] Usate of tmp directories. R has a robust interface for creating temporary files and directories, namely tempfile() and tempdir(). That interface is safer as it (i) can check whether the path already exists and (ii) points to the system temp path, which can be cleared after the R session ends. Currently, the package makes artificial temporary directories manually, for instance, in the faasr() function.

[reply 8] Thank you for the suggestion! FaaSr creates two kinds of directories: faasr_gh_local_repo and faasr_data. faasr_gh_local_repo contains a local repository for github and faasr_data contains files for local testing. Both should be persistent and accessible for users. We’ll be finding a better way to use tempfile() and tempdir().

[comment 9] Use of paste0() for path creation. To ensure cross-os compatibility, it is recommended that paths should be created using the file.path() function. I have added a few examples in a PR I am submitting with this review.

[reply 9] thank you again for the suggestion, and thank you for the PR! this is a very useful hint, and we will implement it  throughout the code - https://github.com/FaaSr/FaaSr-package/issues/114
