### 2024-10-07 #7027 @danielskatz
👋 @gvegayon - Have you had a chance to review and respond to the points above?
