### 2024-10-18 #7027 @gvegayon
Thank you, @renatof and @danielskatz for your patience. Here are my responses to @renatof's; the only missing issue would be point 2, about other R packages:

- reply 1: 👍🏼
- reply 2: I understand your point, yet those two R packages do provide part of the functionality yours does. Although it is not as general, it is part of functions as a service (I understand). I would defer to @danielskatz to whether mentioning these two other packages makes sense or not.
- reply 3: 👍🏼
- reply 4: 👍🏼
- reply 5: 👍🏼
- reply 6: 👍🏼
- reply 7: 👍🏼, nonetheless, here is [one example](https://github.com/FaaSr/FaaSr-package/blob/c26b5edd32bc1cadf403d5bd98a4b3aeb8cced80/R/faasr_client_tools.R#L39-L43).
- reply 8: 👍🏼
- reply 9: 👍🏼
