### 2024-10-24 e17ad14
added lambdr and aws.lambda refs to JOSS paper
