+++ paper/paper.bib
@@ -160,3 +160,21 @@
   howpublished = {[Online], Available: \url{https://github.com/FaaSr/FaaSr-JSON-Builder}},
   note = {[Accessed: Jun-2024]}
 }
+
+@Manual{lambdr,
+    title = {lambdr: Create a Runtime for Serving Containerised R Functions on AWS Lambda},
+    author = {David Neuzerling, James Goldie},
+    year = {2023},
+    note = {lambdr package version 1.2.5},
+    url = {https://CRAN.R-project.org/package=lambdr},
+    doi = {10.32614/CRAN.package.lambdr}
+  }
+
+@Manual{awslambdaR,
+    title = {aws.lambda: AWS Lambda Client Package},
+    author = {Thomas J. Leeper, John Harmon},
+    year = {2020},
+    note = {aws.lambda package version 0.2.0},
+    url = {https://cran.r-project.org/web/packages/aws.lambda},
+    doi = {10.32614/CRAN.package.aws.lambda}
+  }
