### 2024-10-24 9651cab
updated joss paper bib refs
