+++ paper/paper.bib
@@ -163,7 +163,7 @@
 
 @Manual{lambdr,
     title = {lambdr: Create a Runtime for Serving Containerised R Functions on AWS Lambda},
-    author = {David Neuzerling, James Goldie},
+    author = {Neuzerling, David and Goldie, James},
     year = {2023},
     note = {lambdr package version 1.2.5},
     url = {https://CRAN.R-project.org/package=lambdr},
@@ -172,7 +172,7 @@
 
 @Manual{awslambdaR,
     title = {aws.lambda: AWS Lambda Client Package},
-    author = {Thomas J. Leeper, John Harmon},
+    author = {Leeper, Thomas and Harmon, John},
     year = {2020},
     note = {aws.lambda package version 0.2.0},
     url = {https://cran.r-project.org/web/packages/aws.lambda},
