### 2024-10-24 #7027 @danielskatz
@spark0510 and @renatof - I'll now proofread the paper, and let you know next steps.
