### 2024-10-24 78fd3d0
making URLs link in paper.md
