### 2024-10-24 #7027 @danielskatz
@spark0510 and @renatof - I'm suggesting some changes in https://github.com/FaaSr/FaaSr-package/pull/116 - please merge this or let me know what you disagree with. If you want to update the paper.pdf in the repo, please do that too, but JOSS doesn't need it, as we'll soon be generating our own version for the publication.
