### 2024-11-01 #7027 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
