+++ paper/paper.md
@@ -59,7 +59,7 @@ The FaaSr package consists of server-side and client-side functions. The server-
 
 The client-side functions are executed iteratively by a user from their desktop environment (e.g., RStudio). The primary client-side functions exposed to users allow them to: 1) register workflows with FaaS providers, 2) invoke workflows as either a one-off or to set timer schedules for triggering workflows at pre-specified intervals, and 3) copy execution logs from S3 storage to their desktop. The client-side interfaces build on the R  `faasr` function, which creates an object instance in memory in an R session for the user, and which can then be subsequently used to register and invoke functions. This function takes as arguments the name of a JSON-formatted [@pezoa2016foundations] workflow configuration file, and (optionally) the name of a file storing FaaS/S3 cloud provider credentials. The JSON schema for this file is also stored in the FaaSr-Package repository.
 
-FaaSr supports the execution of workflows that can be expressed as a Directed Acyclic Graph (DAG) of functions. The graph (specifying functions and their dependences) is described in JSON format, which can be generated automatically from a Web-based graphical editor using the FaaSr-JSON-Builder tool [@faasr-json-builder] . \autoref{fig:workflow} shows an example workflow DAG graph with ten functions for an ecological forecasting application.
+FaaSr supports the execution of workflows that can be expressed as a Directed Acyclic Graph (DAG) of functions. The graph (specifying functions and their dependences) is described in JSON format, which can be generated automatically from a Web-based graphical editor using the FaaSr-JSON-Builder tool [@faasr-json-builder]. \autoref{fig:workflow} shows an example workflow DAG graph with ten functions for an ecological forecasting application.
 
 ![Fig. 1. FaaSr Example Workflow.\label{fig:workflow}](FaaSr_example_workflow.png)
 
