### 2024-11-01 #7027 @danielskatz
@renatof - I apologize, but I missed one small fix in the paper, as indicated in https://github.com/FaaSr/FaaSr-package/pull/117. If you can merge this, I can go ahead and complete all the remaining steps and publish this.
