### 2024-11-01 #7027 @danielskatz
Congratulations to @spark0510 (Sungjae Park) and @renatof on your publication!!

And thanks to @tyson-swetnam, @yadudoc, and @gvegayon for reviewing!
JOSS depends on volunteers and we wouldn't be able to do this work for the community without you
